moved parse.go to outermost level
