+++ parse.go
@@ -1,17 +1,19 @@
-package parser
+package main
 
 import (
 	"encoding/json"
+
 	"github.com/pkg/errors"
 
-	devfileCtx "github.com/devfile/parser/pkg/devfile/parser/context"
-	"github.com/devfile/parser/pkg/devfile/parser/data"
+	devfile "github.com/devfile/parser/pkg/devfile"
+	devfileCtx "github.com/devfile/parser/pkg/devfile/context"
+	"github.com/devfile/parser/pkg/devfile/data"
 	"github.com/devfile/parser/pkg/devfile/validate"
 )
 
 // ParseDevfile func validates the devfile integrity.
 // Creates devfile context and runtime objects
-func parseDevfile(d DevfileObj) (DevfileObj, error) {
+func parseDevfile(d devfile.DevfileObj) (devfile.DevfileObj, error) {
 
 	// Validate devfile
 	err := d.Ctx.Validate()
@@ -37,7 +39,7 @@ func parseDevfile(d DevfileObj) (DevfileObj, error) {
 
 // Parse func populates the devfile data, parses and validates the devfile integrity.
 // Creates devfile context and runtime objects
-func parse(path string) (d DevfileObj, err error) {
+func parse(path string) (d devfile.DevfileObj, err error) {
 
 	// NewDevfileCtx
 	d.Ctx = devfileCtx.NewDevfileCtx(path)
@@ -54,7 +56,7 @@ func parse(path string) (d DevfileObj, err error) {
 // and validates the devfile integrity with the schema
 // and validates the devfile data.
 // Creates devfile context and runtime objects.
-func ParseAndValidate(path string) (d DevfileObj, err error) {
+func ParseAndValidate(path string) (d devfile.DevfileObj, err error) {
 
 	// read and parse devfile from given path
 	d, err = parse(path)
@@ -74,7 +76,7 @@ func ParseAndValidate(path string) (d DevfileObj, err error) {
 
 // parseInMemory func populates the data from memory, parses and validates the devfile integrity.
 // Creates devfile context and runtime objects
-func parseInMemory(bytes []byte) (d DevfileObj, err error) {
+func parseInMemory(bytes []byte) (d devfile.DevfileObj, err error) {
 
 	// Fill the fields of DevfileCtx struct
 	err = d.Ctx.PopulateFromBytes(bytes)
@@ -88,7 +90,7 @@ func parseInMemory(bytes []byte) (d DevfileObj, err error) {
 // and validates the devfile integrity with the schema
 // and validates the devfile data.
 // Creates devfile context and runtime objects.
-func ParseInMemoryAndValidate(data []byte) (d DevfileObj, err error) {
+func ParseInMemoryAndValidate(data []byte) (d devfile.DevfileObj, err error) {
 
 	// read and parse devfile from given data
 	d, err = parseInMemory(data)
