+++ pkg/devfile/context/apiVersion.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/devfile/parser/pkg/devfile/parser/data"
+	"github.com/devfile/parser/pkg/devfile/data"
 	"github.com/pkg/errors"
 	"k8s.io/klog"
 )
