+++ pkg/devfile/context/schema.go
@@ -3,7 +3,7 @@ package parser
 import (
 	"fmt"
 
-	"github.com/devfile/parser/pkg/devfile/parser/data"
+	"github.com/devfile/parser/pkg/devfile/data"
 	"github.com/pkg/errors"
 	"github.com/xeipuuv/gojsonschema"
 	"k8s.io/klog"
