+++ pkg/devfile/context/schema_test.go
@@ -3,7 +3,7 @@ package parser
 import (
 	"testing"
 
-	v100 "github.com/devfile/parser/pkg/devfile/parser/data/1.0.0"
+	v100 "github.com/devfile/parser/pkg/devfile/data/1.0.0"
 )
 
 const (
