+++ pkg/devfile/data/1.0.0/components.go
@@ -3,7 +3,7 @@ package version100
 import (
 	"strings"
 
-	"github.com/devfile/parser/pkg/devfile/parser/data/common"
+	"github.com/devfile/parser/pkg/devfile/data/common"
 )
 
 func (d *Devfile100) GetMetadata() common.DevfileMetadata {
