+++ pkg/devfile/data/1.0.0/devfileJsonSchema100.go
