+++ pkg/devfile/data/2.0.0/components_test.go
@@ -3,7 +3,7 @@ package version200
 import (
 	"testing"
 
-	common "github.com/devfile/parser/pkg/devfile/parser/data/common"
+	common "github.com/devfile/parser/pkg/devfile/data/common"
 )
 
 func TestGetCommands(t *testing.T) {
