+++ pkg/devfile/data/2.0.0/devfileJsonSchema200.go
