+++ pkg/devfile/data/2.1.0/types.go
@@ -1,6 +1,6 @@
 package version210
 
-import "github.com/devfile/parser/pkg/devfile/parser/data/common"
+import "github.com/devfile/parser/pkg/devfile/data/common"
 
 // CommandGroupType describes the kind of command group.
 // +kubebuilder:validation:Enum=build;run;test;debug
