+++ pkg/devfile/data/helper_test.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"testing"
 
-	v100 "github.com/devfile/parser/pkg/devfile/parser/data/1.0.0"
+	v100 "github.com/devfile/parser/pkg/devfile/data/1.0.0"
 )
 
 func TestNewDevfileData(t *testing.T) {
