+++ pkg/devfile/validate/components.go
@@ -3,7 +3,7 @@ package validate
 import (
 	"fmt"
 
-	"github.com/devfile/parser/pkg/devfile/parser/data/common"
+	"github.com/devfile/parser/pkg/devfile/data/common"
 )
 
 // Errors
