+++ pkg/devfile/validate/validate.go
@@ -3,12 +3,12 @@ package validate
 import (
 	"reflect"
 
-	"github.com/devfile/parser/pkg/devfile/parser/data/common"
+	"github.com/devfile/parser/pkg/devfile/data/common"
 	"k8s.io/klog"
 
-	v100 "github.com/devfile/parser/pkg/devfile/parser/data/1.0.0"
-	v200 "github.com/devfile/parser/pkg/devfile/parser/data/2.0.0"
-	v210 "github.com/devfile/parser/pkg/devfile/parser/data/2.1.0"
+	v100 "github.com/devfile/parser/pkg/devfile/data/1.0.0"
+	v200 "github.com/devfile/parser/pkg/devfile/data/2.0.0"
+	v210 "github.com/devfile/parser/pkg/devfile/data/2.1.0"
 )
 
 // ValidateDevfileData validates whether sections of devfile are odo compatible
