+++ pkg/devfile/writer.go
@@ -1,7 +1,8 @@
-package parser
+package devfile
 
 import (
 	"encoding/json"
+
 	"gopkg.in/yaml.v2"
 
 	"github.com/pkg/errors"
