+++ pkg/testingutil/devfile.go
@@ -1,8 +1,8 @@
 package testingutil
 
 import (
-	"github.com/devfile/parser/pkg/devfile/parser/data/common"
-	versionsCommon "github.com/devfile/parser/pkg/devfile/parser/data/common"
+	"github.com/devfile/parser/pkg/devfile/data/common"
+	versionsCommon "github.com/devfile/parser/pkg/devfile/data/common"
 )
 
 // TestDevfileData is a convenience data type used to mock up a devfile configuration
